Added Sigmoid functions.

+++ lib/ZZNetwork.h
@@ -16,7 +16,7 @@ public:
 
     void train(int **trainingSet, int setSize);
     //void train(double **trainingSet, int nbFeatures, int setSize);
-    int * predict(int *input);
+    double * predict(int *input);
     //Test function
 
     operator bool() const {
